Phase 2 of turnout retry automation - allow parameter settings to be set per turnout, named, and used across multiple turnouts

+++ jmri/jmrit/turnoutoperations/CommonTurnoutOperationConfig.java
@@ -70,7 +70,7 @@ public class CommonTurnoutOperationConfig extends TurnoutOperationConfig {
 	 * called when OK button pressed in config panel, to retrieve and set new values
 	 */
 	public void endConfigure() {
-		long newInterval = ((Integer)intervalSpinner.getValue()).intValue();
+		int newInterval = ((Integer)intervalSpinner.getValue()).intValue();
 		myOp.setInterval(newInterval);
 		int newMaxTries = ((Integer)maxTriesSpinner.getValue()).intValue();
 		myOp.setMaxTries(newMaxTries);
